third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/worker-interception.https-expected.txt

CONSOLE ERROR: line 14: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('https://127.0.0.1:8444') does not match the recipient window's origin ('https://web-platform.test:8444').
CONSOLE ERROR: line 14: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('https://127.0.0.1:8444') does not match the recipient window's origin ('https://web-platform.test:8444').
CONSOLE ERROR: line 14: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('https://127.0.0.1:8444') does not match the recipient window's origin ('https://web-platform.test:8444').
This is a testharness.js-based test.
FAIL Verify worker script from uncontrolled document is intercepted by Service Worker promise_test: Unhandled rejection with value: undefined
FAIL Verify worker script intercepted by same-origin response succeeds promise_test: Unhandled rejection with value: undefined
FAIL Verify worker script intercepted by cors response succeeds promise_test: Unhandled rejection with value: undefined
PASS Verify worker script intercepted by no-cors cross-origin response fails 
PASS Verify worker loads from controlled document are intercepted by Service Worker 
Harness: the test ran to completion.
